refs #180 Fixing issue with the migration on Install.  When installing the Migrations library creates the migration table schema in the new structure so there is no need for this migration.  This migration is only for existing installations.

+++ bonfire/application/db/migrations/core/012_Migration_schema_change.php
@@ -2,6 +2,10 @@
 
 class Migration_Migration_schema_change extends Migration {
 	
+	/**
+	 * This Migration file should not be used in an install situation
+	 * because the migration table schema will already be setup correctly
+	 */
 	public function up() 
 	{
 		$prefix = $this->db->dbprefix;
@@ -11,50 +15,64 @@ class Migration_Migration_schema_change extends Migration {
 		$schema_version_query = $this->db->query($sql);
 		$version_array = $schema_version_query->row_array();
 
-		// backup the schema_version table
-		$this->dbforge->rename_table($prefix.'schema_version', $prefix.'schema_version_old');
-
-		// modify the schema_version table
-		$fields = array(
-						'type' => array(
-							'type' => 'VARCHAR',
-							'constraint' => 20, 
-							'null' => FALSE,
-						),
-						'version_num' => array(
-							'type' => 'INT',
-							'constraint' => '4',
-							'default'    => 0,
-						),
-				);
-		$this->dbforge->add_field($fields);
-		$this->dbforge->add_key('type', TRUE);
-		$this->dbforge->create_table('schema_version');
-		
-		// add records for each of the old permissions
-		foreach ($version_array as $type => $version_num)
+		// check if the table is in the old format
+		if (!isset($version_array->type))
 		{
-			$type_field = $type == 'version' ? 'core' : str_replace('version', '', $type);
-			
-			if ($type_field == 'core')
+			// the table is in the old format
+
+			// backup the schema_version table
+			$this->dbforge->rename_table($prefix.'schema_version', $prefix.'schema_version_old');
+
+			// modify the schema_version table
+			$fields = array(
+							'type' => array(
+								'type' => 'VARCHAR',
+								'constraint' => 20, 
+								'null' => FALSE,
+							),
+							'version_num' => array(
+								'type' => 'INT',
+								'constraint' => '4',
+								'default'    => 0,
+							),
+					);
+			$this->dbforge->add_field($fields);
+			$this->dbforge->add_key('type', TRUE);
+			$this->dbforge->create_table('schema_version');
+
+			// add records for each of the old permissions
+			foreach ($version_array as $type => $version_num)
 			{
-				$version_num++;
+				$type_field = $type == 'version' ? 'core' : str_replace('version', '', $type);
+
+				if ($type_field == 'core')
+				{
+					$version_num++;
+				}
+
+				$this->db->query("INSERT INTO {$prefix}schema_version VALUES ('{$type_field}', ".$version_num.");");
 			}
-			
-			$this->db->query("INSERT INTO {$prefix}schema_version VALUES ('{$type_field}', ".$version_num.");");
 		}
 	}
 	
 	//--------------------------------------------------------------------
 	
+	/**
+	 * This Migration file should not be used in an install situation
+	 * because the migration table schema will already be setup correctly
+	 * and the old table won't exist
+	 */
 	public function down() 
 	{
 		$prefix = $this->db->dbprefix;
 		
-		// Reverse the schema_version table changes
-		$this->dbforge->drop_table('schema_version');
-		
-		$this->dbforge->rename_table($prefix.'schema_version_old', $prefix.'schema_version');
+		// check if the old schema exists
+		if ($this->db->table_exists('schema_version_old')) {
+			// Reverse the schema_version table changes
+			$this->dbforge->drop_table('schema_version');
+
+			$this->dbforge->rename_table($prefix.'schema_version_old', $prefix.'schema_version');
+		}
 
 	}
